requirements.txt
opencv-python
adbutils
pyyaml
numpy
requests==2.31.0
loguru
pydantic-settings